Problem: Windows build gives warning on v2_decoder.cpp:130
Solution: add explicit cast

+++ src/v2_decoder.cpp
@@ -125,9 +125,9 @@ int zmq::v2_decoder_t::size_ready(uint64_t msg_size, unsigned char const* read_p
         // construct message using n bytes from the buffer as storage
         // increase buffer ref count
         // if the message will be a large message, pass a valid refcnt memory location as well
-        rc = in_progress.init( (unsigned char*)read_pos, msg_size,
-                               shared_message_memory_allocator::call_dec_ref, buffer(),
-                               provide_refcnt() );
+        rc = in_progress.init ((unsigned char *) read_pos, static_cast <size_t> (msg_size),
+                                shared_message_memory_allocator::call_dec_ref, buffer(),
+                                provide_refcnt ());
 
         // For small messages, data has been copied and refcount does not have to be increased
         if (in_progress.is_zcmsg())
